挿絵アップロードシステム追加

+++ cake/app/Controller/PageController.php
@@ -86,19 +86,82 @@ class PageController extends AppController {
 	}
 
 	/**
-	 * オーディオブックフォルダを初期化する（羅生門、ごん狐、羅生門体験版は削除する）
+	 * 朗読時に表示する画像をアップロードする。
 	 */
-	public function init() {
+	public function image() {
+		$this->layout = 'admin';
+
 		if($_SERVER['HTTP_HOST'] !== 'roudoku' && $_SERVER['HTTP_HOST'] !== '133.130.59.45') {
 			$this->redirect('index');
 		}
-		$files = $this->getFileList(getcwd().'/audio/'.AUDIO_BOOKS_FOLDER_NAME);
+
+		if($this->request->is('post')) {
+			$this->d($_POST);
+			$this->d($_FILES);
+
+			$event = '';
+			foreach($_FILES['image']['size'] as $key => $file) {
+				if(empty($file)) {
+					continue;
+				}
+				$img_name = $_POST['title']. "_" . sprintf("%03d", $key+1) . '.png';
+				move_uploaded_file($_FILES['image']['tmp_name'][$key],
+						"audio/event/{$img_name}");
+	 			$event .= "evnt[{$key}] = {'start':'', 'end':'', 'img_path':'', 'audio_id':''};\n";
+				$event .= "evnt[{$key}]['start'] = '" . $_POST["image_start"][$key] . "';\n";
+				$event .= "evnt[{$key}]['end'] = '" . $_POST["image_end"][$key] . "';\n";
+				$event .= "evnt[{$key}]['img_path'] = '/audio/event/{$img_name}';\n";
+				$event .= "evnt[{$key}]['audio_id'] = '';\n";
+			}
+			file_put_contents("audio/event/{$_POST['title']}_".AUDIO_BOOKS_FOLDER_NAME."_frame.event", $event, LOCK_EX);
+
+			$this->set('result', true);
+			return true;
+		}
+
+		$files = $this->getFileList("audio/".AUDIO_BOOKS_FOLDER_NAME);
+		$titles = array();
 		foreach($files as $file) {
-			$f = basename($file);
-			$p1 = preg_match("/gongitune/", $f);
-			$p2 = preg_match("/rashomon/", $f);
-			if(!$p1 && !$p2) {
-				unlink($file);
+			if(substr($file, -5, 5) === 'title') {
+				$titles[basename($file, '.title')] = trim(file_get_contents($file));
+			}
+		}
+		$this->set('titles', $titles);
+	}
+
+	/**
+	 * オーディオブックフォルダ及びイベントフォルダを初期化する（羅生門、ごん狐は残す）
+	 */
+	public function init($target = null) {
+		if($_SERVER['HTTP_HOST'] !== 'roudoku' && $_SERVER['HTTP_HOST'] !== '133.130.59.45') {
+			$this->redirect('index');
+		}
+
+		if($target === null) {
+			$this->redirect('/');
+		}
+
+		if($target === 'upload') {
+			$files = $this->getFileList(getcwd().'/audio/'.AUDIO_BOOKS_FOLDER_NAME);
+			foreach($files as $file) {
+				$f = basename($file);
+				$p1 = preg_match("/gongitune/", $f);
+				$p2 = preg_match("/rashomon/", $f);
+				if(!$p1 && !$p2) {
+					unlink($file);
+				}
+			}
+		}
+
+		if($target === 'image') {
+			$files = $this->getFileList(getcwd().'/audio/event');
+			foreach($files as $file) {
+				$f = basename($file);
+				$p1 = preg_match("/gongitune/", $f);
+				$p2 = preg_match("/rashomon/", $f);
+				if(!$p1 && !$p2) {
+					unlink($file);
+				}
 			}
 		}
 	}
